PM-11123: Moved NotificationStatusDetails to Models/Data
