+++ src/Core/NotificationCenter/Queries/Interfaces/IGetNotificationStatusDetailsForUserQuery.cs
@@ -1,5 +1,5 @@
 ﻿#nullable enable
-using Bit.Core.NotificationCenter.Entities;
+using Bit.Core.NotificationCenter.Models.Data;
 using Bit.Core.NotificationCenter.Models.Filter;
 
 namespace Bit.Core.NotificationCenter.Queries.Interfaces;
