+++ test/Core.Test/NotificationCenter/AutoFixture/NotificationStatusDetailsFixtures.cs
@@ -1,6 +1,6 @@
 ﻿#nullable enable
 using AutoFixture;
-using Bit.Core.NotificationCenter.Entities;
+using Bit.Core.NotificationCenter.Models.Data;
 using Bit.Test.Common.AutoFixture.Attributes;
 
 namespace Bit.Core.Test.NotificationCenter.AutoFixture;
@@ -11,8 +11,7 @@ public class NotificationStatusDetailsCustomization : ICustomization
     {
         fixture.Customize<NotificationStatusDetails>(composer => composer.With(n => n.Id, Guid.NewGuid())
             .With(n => n.UserId, Guid.NewGuid())
-            .With(n => n.OrganizationId, Guid.NewGuid())
-            .With(n => n.NotificationStatusUserId, Guid.NewGuid()));
+            .With(n => n.OrganizationId, Guid.NewGuid()));
     }
 }
